style: 라벨 디렉토리 추가 및 styled component로 변경
