+++ github-issue/src/css/Header.css
@@ -1,6 +1,3 @@
-.header-container {
-  padding: 50px 10% 0 10%;
-}
 
 .tab-button {
   height: 36px;
