+++ github-issue/src/css/LabelCustom.css
@@ -1,6 +1,6 @@
 .label-custom-container {
-  margin: 50px 10% 0 10%;
-  padding: 20px;
+  margin-top: 3em;
+  padding: 1.5em;
   border: 1px solid #d9dce0;
   border-radius: 5px;
 }
