+++ github-issue/src/css/LabelList.css
@@ -1,6 +1,3 @@
-.label-list-container {
-  padding: 50px 10% 0 10%;
-}
 
 .label-list {
   padding: 0px;
